requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
ollama==0.1.6
qdrant-client>=1.7.1
pypdf2==3.0.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.2.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
apscheduler==3.10.4
websockets==12.0
